Add more tests for moving category directory

+++ app/models/category.rb
@@ -36,7 +36,7 @@ class Category < ActiveRecord::Base
       @old_path = "#{@old_dir}/#{pdf.filename}"
 
       # Move the category directory to the new one.
-      if File.exists?(pdf.full_dir)
+      if File.exists?(pdf.full_dir) and File.exists?(@old_path)
         File.rename(@old_path, pdf.full_path)
       elsif File.exists?(@old_dir)
         File.rename(@old_dir,pdf.full_dir)
